Added instructions GETI, GETK and improved Clock precision
Improved Clock precision
Added Instructions: GETI, GETK
Fixed processor not stopping properly on errors

Added class ArgumentOptions.Dbl
Added method ArgumentsParser.defineDbl

+++ src/main/java/io/github/hds/pemu/processor/ProcessorConfig.java
@@ -7,7 +7,7 @@ import org.jetbrains.annotations.NotNull;
 public class ProcessorConfig {
     public int bits;
     public int memSize;
-    public int clock;
+    public double clock;
     public @NotNull InstructionSet instructionSet;
 
     public ProcessorConfig() {
@@ -22,11 +22,11 @@ public class ProcessorConfig {
         this(bits, memSize, 1000);
     }
 
-    public ProcessorConfig(int bits, int memSize, int clock) {
+    public ProcessorConfig(int bits, int memSize, double clock) {
         this(bits, memSize, clock, BasicInstructions.BASIC_SET);
     }
 
-    public ProcessorConfig(int bits, int memSize, int clock, @NotNull InstructionSet instructionSet) {
+    public ProcessorConfig(int bits, int memSize, double clock, @NotNull InstructionSet instructionSet) {
         switch (bits) {
             case Word.SizeBit8:
             case Word.SizeBit16:
